update user and profile schema, move modal to component
