+++ models/profile.js
@@ -3,6 +3,9 @@ const Schema = mongoose.Schema;
 
 const profileSchema = new Schema(
   {
+    name: {
+      type: String,
+      required: true,
     email: {
       type: String,
       required: true,
