+++ src/components/ProfileForm/ProfileForm.jsx
@@ -2,210 +2,192 @@ import React, { useState, useEffect, useRef } from "react";
 import { useForm } from '../../hooks/useForm';
 import * as profileService from "../../services/profileService";
 
-export default function ProfileForm(props) {
+export default function ProfileForm( {state, setShowModal, handleSubmit, handleChange} ) {
   const formRef = useRef();
-  const [formInvalid, setValidForm] = useState(true);
-  const [state, handleChange] = useForm({});
 
-  // // hook to check form validity 
-  // useEffect(() => {
-  // formRef.current.checkValidity() ? setValidForm(false) : setValidForm(true);
-  // }, [state]);
 
   // pass form data via submit to handleAddProfile func 
-  async function handleSubmit(e) {
-    e.preventDefault();
-    const updatedProfile = await profileService.update(state)
-    props.setAccount(updatedProfile)
-  }
+  // async function handleSubmit(e) {
+  //   e.preventDefault();
+  //   const updatedProfile = await profileService.update(state)
+  //   props.setAccount(updatedProfile)
+  // }
   
   // const deleteProfile = () => profileService.deleteOne(props.account._id);
 
-  const dummyData = [
-    "Tarrant County, TX",
-    "Dallas County, TX",
-    "Los Angeles County, CA",
-    "New York County, NY"
-  ]
-
   return (
-    <main className="w-5/6 mt-10 mx-auto">
-      {/* FORM START */}
-      <form ref={formRef} onSubmit={handleSubmit}
-        className="space-y-8 divide-y divide-gray-500"
-      >
-        {/* Styling for FORM divs begins */}
-
-        <div className="pt-8 space-y-6 sm:pt-10 sm:space-y-5">
-          {/* For simplicity, each section is seperated by its own empty div */}
-          <div className="space-y-6 sm:space-y-5">
-            {/* TOP SECTION  */}
-            <div>
-              <div className="sm:grid sm:grid-cols-1 sm:gap-4 sm:items-start sm:pt-5">
-                <h1 className="mt-1 sm:mt-0 sm:col-span-2">
-                  Welcome insert name, 
-                </h1>
-                <div className="mt-1 sm:mt-0 sm:col-span-2">
-                  Edit your account details below 
-                </div>
-              </div>
-
-              <div className="sm:grid sm:grid-cols-3 sm:gap-4 sm:items-start  sm:pt-5">
-                <label
-                  htmlFor="email"
-                  className="block text-sm font-medium text-gray-700 sm:mt-px sm:pt-2"
-                >
-                  Email
-                </label>
-                <div className="mt-1 sm:mt-0 sm:col-span-2">
-                  <input
-                    type="text"
-                    name="email"
-                    id="email"
-                    value={state.email}
-                    onChange={handleChange}
-                    className="px-2 max-w-lg block w-full shadow-sm focus:ring-indigo-500 focus:border-indigo-500 sm:max-w-xs sm:text-sm border-gray-300 rounded-md"
-                  />
-                </div>
-              </div>
-
-              <div className="sm:grid sm:grid-cols-3 sm:gap-4 sm:items-start  sm:pt-5">
-                <label
-                  htmlFor="movingFrom"
-                  className="block text-sm font-medium text-gray-700 sm:mt-px sm:pt-2"
-                >
-                  Moving From:
-                </label>
-                <div className="mt-1 sm:mt-0 sm:col-span-2">
-                  <input
-                    type="text"
-                    name="movingFrom"
-                    id="movingFrom"
-                    value={state?.movingFrom}
-                    onChange={handleChange}
-                    className="px-2 max-w-lg block w-full shadow-sm focus:ring-indigo-500 focus:border-indigo-500 sm:max-w-xs sm:text-sm border-gray-300 rounded-md"
-                  />
-                </div>
-              </div>
-            </div>
-            {/*  RECENTLY BROWSED SECTION */}
-            {/* using dummy data and simple stlying to get work started */}
-            <div>
-              <div className="sm:items-start">
-                <div className="sm:grid sm:grid-cols-3 sm:gap-4 sm:items-start sm:border-t sm:border-gray-200 sm:pt-5">
-                  <label
-                    htmlFor="recently-browsed"
-                    className="block text-sm font-medium text-gray-700 sm:mt-px sm:pt-2"
-                  >
-                    Recently Browsed Counties
-                  </label>
-                  
-                </div>
-
-                {/* Map over recentCountries db field for user */}
-                <div className="mt-4 sm:mt-4 sm:col-span-1 flex justify-center ">
-                { dummyData ? dummyData.map((item, key) => {
-                    return (
-                      <div className="py-2 px-2 ml-6 w-1/5 h-28 border border-black" key={key}>
-                        <p className=''>{item}</p>
-                      </div>
-                    );
-                  })
-                : "No recently browsed counties"
-                }  
-                </div>
-              </div>
-            </div>
-            {/* GROUP SECTION  */}
-            {/* need to create list of groups to join with checkbox save feature */}
-            <div>
-              <div className="sm:grid sm:grid-cols-1 sm:gap-4 sm:items-start sm:border-t sm:border-gray-200 sm:pt-5 h-auto">
-                <div>
-                  <h3 className="mt-4 text-lg leading-6 font-medium text-gray-900">
-                    Groups
-                  </h3>
-                  <p className="mt-2 mx-auto max-w-2xl text-sm text-gray-500">
-                    Choose the options that best represents you (This information may be shared publicly)
-                  </p>
-                </div>
-              </div>
-            </div>
-            {/* PREFERENCES SECTION  */}
-            <div>
-            <div className="sm:grid sm:grid-cols-1 sm:gap-4 sm:items-start sm:border-t sm:border-gray-200 sm:pt-5 h-auto">
-              <div>
-                <h3 className="mt-4 text-lg leading-6 font-medium text-gray-900">
-                  Preferences
-                </h3>
-              </div>
-            </div>
-
-            <div className="sm:grid sm:grid-cols-3 sm:gap-4 sm:items-start  sm:pt-5">
-              <label
-                htmlFor="language"
-                className="block text-sm font-medium text-gray-700 sm:mt-px sm:pt-2"
-              >
-                Language
-              </label>
+     <>
+      {/* Modal for Profile content change */}
+      <div class="min-w-screen animated fadeIn faster  fixed  left-0 top-0 flex justify-center items-center inset-0 z-50 outline-none focus:outline-none bg-no-repeat bg-center bg-cover"  id="modal-id">
+        <div class="absolute bg-black opacity-80 inset-0 z-0"></div>
+        <div class="w-full  max-w-lg p-5 relative mx-auto my-auto rounded-xl shadow-lg  bg-white ">
+          {/* <!--content--> */}
+          <div class="">
+            {/* <!--body--> */}
+            <div class="text-center flex-auto justify-center">
+              <h2 class="text-xl font-bold p-2 ">Profile Settings</h2>
               <div className="mt-1 sm:mt-0 sm:col-span-2">
-                  <input
-                        type="text"
-                        name="language"
-                        id="language"
-                        placeholder="English"
-                        value={state?.language}
-                        onChange={handleChange}
-                        className="px-2 max-w-lg block w-full shadow-sm focus:ring-indigo-500 focus:border-indigo-500 sm:max-w-xs sm:text-sm border-gray-300 rounded-md"
+                Edit your account details below 
+              </div>
+              {/* FORM START */}
+              <form ref={formRef} onSubmit={handleSubmit}
+              className=""
+              >
+              {/* Styling for FORM divs begins */}
+
+              <div className="">
+              {/* For simplicity, each section is seperated by its own empty div */}
+              <div className="">
+                  {/* TOP SECTION  */}
+                  <div className="sm:grid sm:grid-cols-3 sm:gap-4 sm:items-start sm:pt-5 mb-2">
+                      <label
+                      htmlFor="name"
+                      className="block text-sm font-medium text-gray-700 "
+                      >
+                      Name:
+                      </label>
+                      <div className="mt-1 sm:mt-0 sm:col-span-2 self-center">
+                      <input
+                          type="text"
+                          name="name"
+                          id="name"
+                          value={state.name}
+                          onChange={handleChange}
+                          className="px-2 max-w-lg block w-full shadow-sm focus:ring-indigo-500 focus:border-indigo-500 sm:max-w-xs sm:text-sm border-gray-300 rounded-md"
                       />
-              </div>
-            </div>
+                      </div>
+                  </div>
+                  <div className="sm:grid sm:grid-cols-3 sm:gap-4 sm:items-start mb-2 ">
+                      <label
+                      htmlFor="email"
+                      className="block text-sm font-medium text-gray-700"
+                      >
+                      Email
+                      </label>
+                      <div className="mt-1 sm:mt-0 sm:col-span-2 self-center">
+                      <input
+                          type="text"
+                          name="email"
+                          id="email"
+                          value={state.email}
+                          onChange={handleChange}
+                          className="px-2 max-w-lg block w-full shadow-sm focus:ring-indigo-500 focus:border-indigo-500 sm:max-w-xs sm:text-sm border-gray-300 rounded-md"
+                      />
+                      </div>
+                  </div>
 
-            <div className="sm:grid sm:grid-cols-3 sm:gap-4 sm:items-start sm:border-t sm:border-gray-200 sm:pt-5">
-              <label
-                htmlFor="country"
-                className="block text-sm font-medium text-gray-700 sm:mt-px sm:pt-2"
-              >
-                Date Format
-              </label>
-              <div className="mt-1 sm:mt-0 sm:col-span-2">
-                <select
-                  id="dateFormat"
-                  name="dateFormat"
-                  value={state.dateFormat}
-                  onChange={handleChange}
-                  className="max-w-lg block focus:ring-indigo-500 focus:border-indigo-500 w-full shadow-sm sm:max-w-xs sm:text-sm border-gray-300 rounded-md"
-                >
-                  <option value='DD-MM-YYYY'>DD-MM-YYYY</option>
-                  <option value='MM-DD-YYYY'>MM-DD-YYYY</option>
-                </select>
+                  <div className="sm:grid sm:grid-cols-3 sm:gap-4 sm:items-start mb-2">
+                      <label
+                      htmlFor="movingFrom"
+                      className="block text-sm font-medium text-gray-700"
+                      >
+                      Moving From:
+                      </label>
+                      <div className="mt-1 sm:mt-0 sm:col-span-2 self-center">
+                      <input
+                          type="text"
+                          name="movingFrom"
+                          id="movingFrom"
+                          value={state?.movingFrom}
+                          onChange={handleChange}
+                          className="px-2 max-w-lg block w-full shadow-sm focus:ring-indigo-500 focus:border-indigo-500 sm:max-w-xs sm:text-sm border-gray-300 rounded-md"
+                      />
+                      
+                      </div>
+                  </div>
+                  <div className="sm:grid sm:grid-cols-3 sm:gap-4 sm:items-start mb-2">
+                      <label
+                      htmlFor="photo"
+                      className="block text-sm font-medium text-gray-700"
+                      >
+                      Profile Photo:
+                      </label>
+                      <div className="mt-1 sm:mt-0 sm:col-span-2 self-center">
+                      <input
+                          type="file"
+                          name="photo"
+                          id="photo"
+                          value={state.photo}
+                          onChange={handleChange}
+                          className="max-w-lg block w-full focus:ring-indigo-500 focus:border-indigo-500 sm:max-w-xs sm:text-sm border-none"
+                      />
+                      </div>
+                  </div>
+                  {/* GROUP SECTION  */}
+                  {/* need to create list of groups to join with checkbox save feature */}
+                  <div>
+                  
+                  </div>
+                  {/* PREFERENCES SECTION  */}
+                  <div>
+                  <div className="sm:grid sm:grid-cols-1 sm:gap-4 sm:items-start sm:border-t sm:border-gray-200 sm:pt-2 h-auto">
+                  <div>
+                      <h3 className="mt-2 text-lg leading-6 font-medium text-gray-900">
+                      Preferences
+                      </h3>
+                  </div>
+                  </div>
+
+                  <div className="sm:grid sm:grid-cols-3 sm:gap-4 sm:items-start  sm:pt-5">
+                    <label
+                        htmlFor="language"
+                        className="block text-sm font-medium text-gray-700"
+                    >
+                        Language
+                    </label>
+                  <div className="mt-1 sm:mt-0 sm:col-span-2">
+                    <select
+                          type="text"
+                          name="language"
+                          id="language"
+                          value={state?.language}
+                          onChange={handleChange}>
+                      <option value="english"> English </option>
+                      <option value="arabic"> Arabic </option>
+                      <option value="chinese"> Chinese </option>
+                      <option value="japanese"> Japanese</option>
+                      <option value="portugese"> Portuguese </option>
+                      <option value="spanish"> Spanish </option>
+                    </select>
+                  </div>
+                  <label
+                      htmlFor="country"
+                      className="block text-sm font-medium text-gray-700 "
+                  >
+                      Date Format
+                  </label>
+                  <div className="mt-1 sm:mt-0 sm:col-span-2">
+                      <select
+                      id="dateFormat"
+                      name="dateFormat"
+                      value={state.dateFormat}
+                      onChange={handleChange}
+                      className="max-w-lg block focus:ring-indigo-500 focus:border-indigo-500 w-full shadow-sm sm:max-w-xs sm:text-sm "
+                      >
+                      <option value='DD-MM-YYYY'>DD-MM-YYYY</option>
+                      <option value='MM-DD-YYYY'>MM-DD-YYYY</option>
+                      </select>
+                  </div>
+                  </div>   
+                  </div>
+                </div>
               </div>
-            </div>   
+              {/* Styling for FORM divs ends */}
+            </form>
+              {/* FORM END */}
+            </div>
+            {/* <!--footer--> */}
+            <div class="p-3  mt-2 text-center space-x-4 md:block">
+                <button class="mb-2 md:mb-0 bg-red-600 px-5 py-2 text-sm shadow-sm font-medium tracking-wider border text-white rounded-full hover:shadow-lg hover:bg-red-900" type="button"
+                    onClick={() => setShowModal(false)}>
+                    Cancel
+                </button>
+                <button class="mb-2 md:mb-0 bg-primary border border-primary px-5 py-2 text-sm shadow-sm font-medium tracking-wider text-white rounded-full hover:shadow-lg hover:bg-yolk hover:text-black" type="button"
+                    onClick={() => setShowModal(false)}>Save</button>
             </div>
           </div>
         </div>
-        {/*  BUTTON SECTION  */}
-        <div>
-          <div className="pt-5">
-            <div className="flex justify-center mb-10">
-              <button
-                type="button"
-                className="bg-white py-2 px-4 border border-gray-300 rounded-md shadow-sm text-sm font-medium text-gray-700 hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
-              >
-                Cancel
-              </button>
-              <button
-                type="submit"
-                disabled={formInvalid}
-                className="ml-3 inline-flex justify-center py-2 px-4 border border-transparent shadow-sm text-sm font-medium rounded-md text-white bg-indigo-600 hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
-              >
-                Save
-              </button>
-            </div>
-          </div>
-        </div>
-        {/* Styling for FORM divs ends */}
-      </form>
-      {/* FORM END */}
-    </main>
+      </div>
+    </>
   );
 }
