+++ src/pages/UserProfile/UserProfile.jsx
@@ -1,180 +1,150 @@
 import React, { useState, useEffect, useRef } from "react";
-import { useHistory } from "react-router-dom";
-import { useForm } from "../../hooks/useForm";
+import { useForm } from '../../hooks/useForm';
 import * as profileService from "../../services/profileService";
 
-import "./UserProfile.styles.css";
+export default function ProfileForm(props) {
+  const formRef = useRef();
+  const [formInvalid, setValidForm] = useState(true);
+  const [state, handleChange] = useForm({});
 
-export default function UserProfile(props) {
-    //  allow us history access for routing
-    const history = useHistory();
-    // initialize form as invalid
-    const [formInvalid, setValidForm] = useState(true);
-    // initialize object for form validation
-    const formRef = useRef();
-    //  custom hook to initialize state
-    const [state, handleChange] = useForm({
-      email: "",
-      movingFrom: "", 
-      language: "",
-      recentCounties: [],
-      dateFormat: "DD-MM-YYYY",
-      groups: [],
-    });
+
+  // pass form data via submit to handleAddProfile func 
+  async function handleSubmit(e) {
+    e.preventDefault();
+    const updatedProfile = await profileService.update(state)
+    props.setAccount(updatedProfile)
+  }
   
-    // function to handle profile create via api
-    async function handleAddProfile(newProfileData) {
-      const newProfile = await profileService.create(newProfileData);
-      console.log(newProfile)
-      history.push("/");
-    }
-    
-    // function to determine if user has profile already, if yes, redirect to edit page
-    // useEffect(() => {
-    //   const hasProfile = async () => {
-    //     const hasData = await profileService.getAllByCurrentUser(props.user._id)
-    //     if (hasData?.name)history.push('/account/edit')
-    //   } 
-    //   hasProfile()
-    // }, []);
-  
-    // hook to check form validity
-    // useEffect(() => {
-    //   formRef.current.checkValidity() ? setValidForm(false) : setValidForm(true);
-    // }, [state]);
-  
-    // pass form data via submit to handleAddprofile func
-    async function handleSubmit(e) {
-      e.preventDefault();
-      handleAddProfile(state);
-    }
-  
-    const dummyData = [
-      "Tarrant County, TX",
-      "Dallas County, TX",
-      "Los Angeles County, CA",
-      "New York County, NY"
-    ]
-    return (
-       <div>
-        {/* FORM START */}
-        <form ref={formRef} onSubmit={handleSubmit}
-        className=""
-        >
+  // const deleteProfile = () => profileService.deleteOne(props.account._id);
+
+  const dummyData = [
+    "Tarrant County, TX",
+    "Dallas County, TX",
+    "Los Angeles County, CA",
+    "New York County, NY"
+  ]
+
+  return (
+    <main className="w-5/6 mt-10 mx-auto">
+      {/* FORM START */}
+      <form ref={formRef} onSubmit={handleSubmit}
+              className=""
+      >
         {/* Styling for FORM divs begins */}
 
-        <div className="">
-        {/* For simplicity, each section is seperated by its own empty div */}
-        <div className="">
+        <div className="pt-8 space-y-6 sm:pt-10 sm:space-y-5">
+          {/* For simplicity, each section is seperated by its own empty div */}
+          <div className="space-y-6 sm:space-y-5">
             {/* TOP SECTION  */}
-            <div className="sm:grid sm:grid-cols-1 sm:gap-4 sm:items-start sm:pt-5">
+            <div>
+              <div className="sm:grid sm:grid-cols-1 sm:gap-4 sm:items-start sm:pt-5">
                 <h1 className="mt-1 sm:mt-0 sm:col-span-2">
-                Welcome insert name, 
+                  Welcome insert name, 
                 </h1>
                 <div className="mt-1 sm:mt-0 sm:col-span-2">
-                Edit your account details below 
+                  Edit your account details below 
                 </div>
-            </div>
+              </div>
 
-            <div className="sm:grid sm:grid-cols-3 sm:gap-4 sm:items-start  sm:pt-5">
+              <div className="sm:grid sm:grid-cols-3 sm:gap-4 sm:items-start  sm:pt-5">
                 <label
-                htmlFor="email"
-                className="block text-sm font-medium text-gray-700 sm:mt-px sm:pt-2"
+                  htmlFor="email"
+                  className="block text-sm font-medium text-gray-700 sm:mt-px sm:pt-2"
                 >
-                Email
+                  Email
                 </label>
                 <div className="mt-1 sm:mt-0 sm:col-span-2">
-                <input
+                  <input
                     type="text"
                     name="email"
                     id="email"
                     value={state.email}
                     onChange={handleChange}
                     className="px-2 max-w-lg block w-full shadow-sm focus:ring-indigo-500 focus:border-indigo-500 sm:max-w-xs sm:text-sm border-gray-300 rounded-md"
-                />
+                  />
                 </div>
-            </div>
+              </div>
 
-            <div className="sm:grid sm:grid-cols-3 sm:gap-4 sm:items-start  sm:pt-5">
+              <div className="sm:grid sm:grid-cols-3 sm:gap-4 sm:items-start  sm:pt-5">
                 <label
-                htmlFor="movingFrom"
-                className="block text-sm font-medium text-gray-700 sm:mt-px sm:pt-2"
+                  htmlFor="movingFrom"
+                  className="block text-sm font-medium text-gray-700 sm:mt-px sm:pt-2"
                 >
-                Moving From:
+                  Moving From:
                 </label>
                 <div className="mt-1 sm:mt-0 sm:col-span-2">
-                <input
+                  <input
                     type="text"
                     name="movingFrom"
                     id="movingFrom"
                     value={state?.movingFrom}
                     onChange={handleChange}
                     className="px-2 max-w-lg block w-full shadow-sm focus:ring-indigo-500 focus:border-indigo-500 sm:max-w-xs sm:text-sm border-gray-300 rounded-md"
-                />
+                  />
                 </div>
+              </div>
             </div>
-            
             {/*  RECENTLY BROWSED SECTION */}
             {/* using dummy data and simple stlying to get work started */}
             <div>
-            <div className="sm:items-start">
+              <div className="sm:items-start">
                 <div className="sm:grid sm:grid-cols-3 sm:gap-4 sm:items-start sm:border-t sm:border-gray-200 sm:pt-5">
-                <label
+                  <label
                     htmlFor="recently-browsed"
-                    className="block text-24 font-medium text-gray-700 sm:mt-px sm:pt-2"
-                >
-                    Recently Browsed 
-                </label>
-                
+                    className="block text-sm font-medium text-gray-700 sm:mt-px sm:pt-2"
+                  >
+                    Recently Browsed Counties
+                  </label>
+                  
                 </div>
 
                 {/* Map over recentCountries db field for user */}
-                <div className="mt-4 sm:mt-4 sm:col-span-1 flex justify-center">
+                <div className="mt-4 sm:mt-4 sm:col-span-1 flex justify-center ">
                 { dummyData ? dummyData.map((item, key) => {
                     return (
-                    <div className="py-2 px-2 ml-6 w-1/5 h-28 border border-black hover:underline" key={key}>
-                        <p className='mt-20'>{item}</p>
-                    </div>
+                      <div className="py-2 px-2 ml-6 w-1/5 h-28 border border-black" key={key}>
+                        <p className=''>{item}</p>
+                      </div>
                     );
-                })
-                : "Search for an ideal area HERE."
+                  })
+                : "No recently browsed counties"
                 }  
                 </div>
-            </div>
+              </div>
             </div>
             {/* GROUP SECTION  */}
             {/* need to create list of groups to join with checkbox save feature */}
             <div>
-            <div className="sm:grid sm:grid-cols-1 sm:gap-4 sm:items-start sm:border-t sm:border-gray-200 sm:pt-5 h-auto">
+              <div className="sm:grid sm:grid-cols-1 sm:gap-4 sm:items-start sm:border-t sm:border-gray-200 sm:pt-5 h-auto">
                 <div>
-                <h3 className="mt-4 text-lg leading-6 font-medium text-gray-900">
+                  <h3 className="mt-4 text-lg leading-6 font-medium text-gray-900">
                     Groups
-                </h3>
-                <p className="mt-2 mx-auto max-w-2xl text-sm text-gray-500">
+                  </h3>
+                  <p className="mt-2 mx-auto max-w-2xl text-sm text-gray-500">
                     Choose the options that best represents you (This information may be shared publicly)
-                </p>
+                  </p>
                 </div>
-            </div>
+              </div>
             </div>
             {/* PREFERENCES SECTION  */}
             <div>
             <div className="sm:grid sm:grid-cols-1 sm:gap-4 sm:items-start sm:border-t sm:border-gray-200 sm:pt-5 h-auto">
-            <div>
+              <div>
                 <h3 className="mt-4 text-lg leading-6 font-medium text-gray-900">
-                Preferences
+                  Preferences
                 </h3>
-            </div>
+              </div>
             </div>
 
             <div className="sm:grid sm:grid-cols-3 sm:gap-4 sm:items-start  sm:pt-5">
-            <label
+              <label
                 htmlFor="language"
                 className="block text-sm font-medium text-gray-700 sm:mt-px sm:pt-2"
-            >
+              >
                 Language
-            </label>
-            <div className="mt-1 sm:mt-0 sm:col-span-2">
-                <input
+              </label>
+              <div className="mt-1 sm:mt-0 sm:col-span-2">
+                  <input
                         type="text"
                         name="language"
                         id="language"
@@ -182,55 +152,56 @@ export default function UserProfile(props) {
                         value={state?.language}
                         onChange={handleChange}
                         className="px-2 max-w-lg block w-full shadow-sm focus:ring-indigo-500 focus:border-indigo-500 sm:max-w-xs sm:text-sm border-gray-300 rounded-md"
-                    />
-            </div>
+                      />
+              </div>
             </div>
 
             <div className="sm:grid sm:grid-cols-3 sm:gap-4 sm:items-start sm:border-t sm:border-gray-200 sm:pt-5">
-            <label
+              <label
                 htmlFor="country"
                 className="block text-sm font-medium text-gray-700 sm:mt-px sm:pt-2"
-            >
+              >
                 Date Format
-            </label>
-            <div className="mt-1 sm:mt-0 sm:col-span-2">
+              </label>
+              <div className="mt-1 sm:mt-0 sm:col-span-2">
                 <select
-                id="dateFormat"
-                name="dateFormat"
-                value={state.dateFormat}
-                onChange={handleChange}
-                className="max-w-lg block focus:ring-indigo-500 focus:border-indigo-500 w-full shadow-sm sm:max-w-xs sm:text-sm border-gray-300 rounded-md"
+                  id="dateFormat"
+                  name="dateFormat"
+                  value={state.dateFormat}
+                  onChange={handleChange}
+                  className="max-w-lg block focus:ring-indigo-500 focus:border-indigo-500 w-full shadow-sm sm:max-w-xs sm:text-sm border-gray-300 rounded-md"
                 >
-                <option value='DD-MM-YYYY'>DD-MM-YYYY</option>
-                <option value='MM-DD-YYYY'>MM-DD-YYYY</option>
+                  <option value='DD-MM-YYYY'>DD-MM-YYYY</option>
+                  <option value='MM-DD-YYYY'>MM-DD-YYYY</option>
                 </select>
-            </div>
+              </div>
             </div>   
             </div>
-        </div>
+          </div>
         </div>
         {/*  BUTTON SECTION  */}
         <div>
-        <div className="pt-5">
+          <div className="pt-5">
             <div className="flex justify-center mb-10">
-            <button
+              <button
                 type="button"
                 className="bg-white py-2 px-4 border border-gray-300 rounded-md shadow-sm text-sm font-medium text-gray-700 hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
-            >
+              >
                 Cancel
-            </button>
-            <button
+              </button>
+              <button
                 type="submit"
                 disabled={formInvalid}
                 className="ml-3 inline-flex justify-center py-2 px-4 border border-transparent shadow-sm text-sm font-medium rounded-md text-white bg-indigo-600 hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500"
-            >
+              >
                 Save
-            </button>
+              </button>
             </div>
-        </div>
+          </div>
         </div>
         {/* Styling for FORM divs ends */}
-        </form>
-        {/* FORM END */}
-    </div>
-    )}
+      </form>
+      {/* FORM END */}
+    </main>
+  );
+}
